Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -14,7 +14,7 @@ int main(){
         if(b==0){
             printf("error");
         }else{
-            printf("%d", a/b);
+            printf("%d", (float)a/b);
         }
     }
     return 0;
